+++ course_overviews/apps.py
@@ -1,0 +1,10 @@
+"""
+Mock course_overviews app configuration.
+"""
+
+from django.apps import AppConfig
+
+
+class CourseOverviewsConfig(AppConfig):
+    name = 'course_overviews'
+    label = 'course_overviews'